+++ field.hh
@@ -34,7 +34,7 @@ public:
 
     double to_double();
 
-    int64_t to_int64();
+    long long to_int64();
 
     /** Return field returned length */
     int      getDataLength() { return datalength; }
